backend/requirements.txt
dj-rest-auth==2.2.5
Django==3.2.12
django-allauth==0.51.0
django-clone==3.0.1
django-countries==7.4.2
django-cors-headers==3.13.0
django-hashedfilenamestorage==2.4.1
django-resized==1.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
PyJWT==2.4.0
requests==2.28.1
requests-oauthlib==1.3.1
